Async video
Summary:
Known issues:
* Still no auth
* Doesn't utilize the `s3://` urls
* Light on tests :(  https://app.asana.com/0/12387442313872/136954532331291

Reviewers: samuel, kunaljham

Reviewed By: kunaljham

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D759

+++ libs/media/audio_service.go
@@ -15,7 +15,7 @@ import (
 type AudioMeta struct {
 	MimeType string
 	Duration time.Duration
-	Size     int
+	Size     uint64
 }
 
 // AudioService implements a media storage service.
@@ -59,7 +59,7 @@ func (s *AudioService) PutReader(id string, r io.ReadSeeker, contentType string)
 	meta := &AudioMeta{
 		MimeType: contentType,
 		Duration: duration,
-		Size:     int(size),
+		Size:     uint64(size),
 	}
 	return meta, errors.Trace(err)
 }
