+++ libs/media/video_service.go
@@ -15,7 +15,7 @@ import (
 type VideoMeta struct {
 	MimeType string
 	Duration time.Duration
-	Size     int
+	Size     uint64
 }
 
 // VideoService implements a media storage service.
@@ -59,7 +59,7 @@ func (s *VideoService) PutReader(id string, r io.ReadSeeker, contentType string)
 	meta := &VideoMeta{
 		MimeType: contentType,
 		Duration: duration,
-		Size:     int(size),
+		Size:     uint64(size),
 	}
 	return meta, errors.Trace(err)
 }
